PyTorch/01-BostonHousing/requirements.txt
# Boston Housing 価格予測プロジェクト - 依存ライブラリ
# Python 3.8以上を推奨

# 深層学習フレームワーク
torch>=2.0.0
torchvision>=0.15.0

# 数値計算・データ処理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# 可視化
matplotlib>=3.7.0
seaborn>=0.12.0

# 設定管理
pyyaml>=6.0

# プログレスバー
tqdm>=4.65.0

# ノートブック（オプション）
jupyter>=1.0.0
ipywidgets>=8.0.0

# デモアプリ（オプション）
# streamlit>=1.28.0
# gradio>=3.50.0
